src/main/java/com/stubulika/resource/React.java
package com.stubulika.resource;

import jdk.nashorn.api.scripting.NashornScriptEngine;

import javax.script.ScriptEngineManager;
import javax.script.ScriptException;
import java.io.InputStream;
import java.io.InputStreamReader;
import java.io.Reader;
import java.util.List;

public class React {

    private ThreadLocal<NashornScriptEngine> engineHolder = new ThreadLocal<NashornScriptEngine>() {
        @Override
        protected NashornScriptEngine initialValue() {
            NashornScriptEngine nashornScriptEngine = (NashornScriptEngine) new ScriptEngineManager().getEngineByName("nashorn");
            try {
                nashornScriptEngine.eval(read("static/admin/nashorn-polyfill.js"));
                nashornScriptEngine.eval(read("static/admin/vendor/react.js"));
                nashornScriptEngine.eval(read("static/admin/vendor/showdown.min.js"));
                nashornScriptEngine.eval(read("static/admin/stub.js"));
            } catch (ScriptException e) {
                throw new RuntimeException(e);
            }
            return nashornScriptEngine;
        }
    };

    public  String renderStubBox(List<StubWrapper> stubs) {
        try {
            Object html = engineHolder.get().invokeFunction("renderServer", stubs);
            return String.valueOf(html);
        }
        catch (Exception e) {
            throw new IllegalStateException("failed to render react component", e);
        }
    }

    private Reader read(String path) {
        InputStream in = getClass().getClassLoader().getResourceAsStream(path);
        return new InputStreamReader(in);
    }
}